+++ cursach/DoctorService.cpp
@@ -25,7 +25,7 @@ void DoctorService::deleteDoctorById(int id) {
 }
 
 Doctor* DoctorService::findDoctorById(int id) {
-	for (int i = 0; i < doctors->size(); i++) {
+	for (size_t i = 0; i < doctors->size(); i++) {
 		if (doctors->at(i)->getId() == id) {
 			return doctors->at(i);
 		}
